[refact/#5]: 컨트롤러 수정
