+++ src/main/java/com/wekids/backend/account/controller/AccountController.java
@@ -1,27 +1,15 @@
 package com.wekids.backend.account.controller;
 
-import com.wekids.backend.account.dto.response.AccountResponse;
-import com.wekids.backend.account.dto.response.BaasAccountResponse;
-import com.wekids.backend.account.service.AccountService;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.client.RestTemplate;
 
-import java.util.List;
+
+
 
 @RestController
 @RequestMapping("/api/v1/accounts")
 @RequiredArgsConstructor
 public class AccountController {
 
-    @Autowired
-    private AccountService accountService;
 
-    @GetMapping()
-    public ResponseEntity<List<AccountResponse>> getAccount(){
-        List<AccountResponse> responseList = accountService.getAccount();
-        return ResponseEntity.ok(responseList);
-    }
 }
